examples/taist_mar_28/line_bot/requirements.txt
fastapi
uvicorn
line-bot-sdk
jinja2
pymongo